+++ tests/module_execution.py
@@ -41,7 +41,7 @@ data_test = [
     # Output interfaces
     (
         spark.nn.interfaces.ExponentialIntegrator, 
-        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(10,) < 0.5, dtype=jnp.float16)),}, 
+        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(10,) < 0.5)),}, 
         {'num_outputs':2,}
     ),
     # Control interfaces
@@ -63,43 +63,43 @@ data_test = [
     # Delays
     (
         spark.nn.delays.NDelays, 
-        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,4) < 0.5, dtype=jnp.float16)),}, 
+        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,4) < 0.5)),}, 
         {'max_delay':5,}
     ),
     (
         spark.nn.delays.N2NDelays, 
-        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,4) < 0.5, dtype=jnp.float16)),}, 
+        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,4) < 0.5)),}, 
         {'units':(2,2), 'max_delay':5,}
     ),
     # Synapses
     (
         spark.nn.synapses.LinearSynapses, 
-        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),}, 
+        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),}, 
         {'units':(2,3),}
     ),
     (
         spark.nn.synapses.LinearSynapses, 
-        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5, dtype=jnp.float16), async_spikes=True),}, 
+        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5), async_spikes=True),}, 
         {'units':(2,3),}
     ),
     (
         spark.nn.synapses.TracedSynapses, 
-        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),}, 
+        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),}, 
         {'units':(2,3),}
     ),
     (
         spark.nn.synapses.TracedSynapses, 
-        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5, dtype=jnp.float16), async_spikes=True),}, 
+        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5), async_spikes=True),}, 
         {'units':(2,3),}
     ),
     (
         spark.nn.synapses.DoubleTracedSynapses, 
-        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),}, 
+        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),}, 
         {'units':(2,3),}
     ),
     (
         spark.nn.synapses.DoubleTracedSynapses, 
-        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5, dtype=jnp.float16), async_spikes=True),}, 
+        {'spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5), async_spikes=True),}, 
         {'units':(2,3),}
     ),
     # Somas
@@ -142,8 +142,8 @@ data_test = [
     (
         spark.nn.learning_rules.HebbianRule, 
         {
-            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),
-            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5, dtype=jnp.float16)),
+            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),
+            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5)),
             'current_kernel': spark.FloatArray(jnp.array(np.random.rand(2,3,4,5), dtype=jnp.float16)),
         }, 
         {'units':(2,3),}
@@ -151,8 +151,8 @@ data_test = [
     (
         spark.nn.learning_rules.HebbianRule, 
         {
-            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5, dtype=jnp.float16), async_spikes=True),
-            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5, dtype=jnp.float16)),
+            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5), async_spikes=True),
+            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5)),
             'current_kernel': spark.FloatArray(jnp.array(np.random.rand(2,3,4,5), dtype=jnp.float16)),
         }, 
         {'units':(2,3),}
@@ -160,8 +160,8 @@ data_test = [
     (
     spark.nn.learning_rules.OjaRule, 
         {
-            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),
-            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5, dtype=jnp.float16)),
+            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),
+            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5)),
             'current_kernel': spark.FloatArray(jnp.array(np.random.rand(2,3,4,5), dtype=jnp.float16)),
         }, 
         {'units':(2,3),}
@@ -169,8 +169,8 @@ data_test = [
     (
     spark.nn.learning_rules.OjaRule, 
         {
-            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5, dtype=jnp.float16), async_spikes=True),
-            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5, dtype=jnp.float16)),
+            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5), async_spikes=True),
+            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5)),
             'current_kernel': spark.FloatArray(jnp.array(np.random.rand(2,3,4,5), dtype=jnp.float16)),
         }, 
         {'units':(2,3),}
@@ -178,8 +178,8 @@ data_test = [
     (
     spark.nn.learning_rules.ZenkeRule, 
         {
-            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),
-            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5, dtype=jnp.float16)),
+            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),
+            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5)),
             'current_kernel': spark.FloatArray(jnp.array(np.random.rand(2,3,4,5), dtype=jnp.float16)),
         }, 
         {'units':(2,3),}
@@ -187,8 +187,8 @@ data_test = [
     (
     spark.nn.learning_rules.ZenkeRule, 
         {
-            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5, dtype=jnp.float16), async_spikes=True),
-            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5, dtype=jnp.float16)),
+            'pre_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3,4,5) < 0.5), async_spikes=True),
+            'post_spikes': spark.SpikeArray(jnp.array(np.random.rand(2,3) < 0.5)),
             'current_kernel': spark.FloatArray(jnp.array(np.random.rand(2,3,4,5), dtype=jnp.float16)),
         }, 
         {'units':(2,3),}
@@ -196,22 +196,22 @@ data_test = [
     # Neurons
     (
         spark.nn.neurons.LIFNeuron, 
-        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),}, 
+        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),}, 
         {'_s_units':(2,3),}
     ),
     (
         spark.nn.neurons.LIFNeuron, 
-        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),}, 
+        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),}, 
         {'_s_units':(2,3),}
     ),
     (
         spark.nn.neurons.ALIFNeuron, 
-        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),}, 
+        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),}, 
         {'_s_units':(2,3),}
     ),
     (
         spark.nn.neurons.ALIFNeuron, 
-        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5, dtype=jnp.float16), async_spikes=False),}, 
+        {'in_spikes': spark.SpikeArray(jnp.array(np.random.rand(4,5) < 0.5), async_spikes=False),}, 
         {'_s_units':(2,3),}
     ),
 ]
